FortuneTeller/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "newsapi-python>=0.2.7",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "pandas-ta>=0.3.14b0",
    "plotly>=6.0.0",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "yfinance>=0.2.52",
]
